+++ tsconfig.json
@@ -4,7 +4,8 @@
       "module": "commonjs",
       "declaration": true,
       "outDir": "./lib",
-      "strict": true
+      "strict": true,
+      "esModuleInterop": true,
     },
     "include": ["src"],
     "exclude": ["node_modules", "**/__tests__/*"]
